Buildpack comes with summon and conjur-api-go pre-installed to speed up retrieval of secrets (#8)
## What does this PR do?

PROBLEM: it's somewhat painful to wait for summon and summon-conjur to install.
SOLUTION: a single binary is a much better proposition. the intention is for each release to carry a conjur-env binary in the vendor directory that can be used with or without the buildpack.
this makes the whole process of secret injection significantly faster i.e. less than a second

The actual buildpack as a tarball can be incorporated as part of the release, it's folder structure looks like so:

```
.
.
├── bin
│   ├── compile
│   ├── decorate
│   └── detect
├── buildpack_vars.sh
├── lib
│   └── 0001_retrieve-secrets.sh
├── upload.sh
└── vendor
    └── conjur-env
```

running `./upload.sh` will upload the buildpack to your cf instance (same flow as before but no need to build)

Interestingly, this also makes it easier for usage with custom buildpacks i.e. download the latest release, and run `./bin/compile <PATH_TO_YOUR_APP>` and you're set :)

## Notes

It's worth inspecting the artefacts over at https://jenkins.conjur.net/job/cyberark--cloudfoundry-conjur-buildpack/job/conjur-env/ to see for yourself.

## Suggestions for improvement

It would be a good idea to add a SHA to each release so users can ensure the binary is legit.